:truck: Move gui related files into new gui folder
